refactor: address PR #3 review feedback - types and constants
- Create internal/types package for shared SecurityCapability type
- Update config.go to use types.SecurityCapability instead of int
- Remove type casting in p2p/downloads.go and p2p/uploads.go
- Consolidate tray.go constants with winservice package
- Update SERVICE.md backup/restore examples with concrete paths
- Update PR3_REVIEW_COMMENTS.md tracking file with completion status
