+++ internal/config/config.go
@@ -4,6 +4,8 @@ import (
 	"os"
 	"strconv"
 	"time"
+
+	"pinshare/internal/types"
 )
 
 // Environment variable names for configuration
@@ -69,7 +71,7 @@ const (
 // AppConfig holds all configuration for the application.
 type AppConfig struct {
 	Version                   string
-	SecurityCapability        int
+	SecurityCapability        types.SecurityCapability
 	UploadFolder              string
 	CacheFolder               string
 	RejectFolder              string
@@ -100,7 +102,7 @@ type AppConfig struct {
 func LoadConfig() (*AppConfig, error) {
 	conf := &AppConfig{
 		Version:                   "dev0.1.3",
-		SecurityCapability:        0,
+		SecurityCapability:        types.SecurityCapabilityNone,
 		UploadFolder:              defaultUploadFolder,
 		CacheFolder:               defaultCacheFolder,
 		RejectFolder:              defaultRejectFolder,
